# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ecommerce Platform API",
    "version": "1.0.0",
    "description": "REST API for the ecommerce platform catalog, auth, health, and uploads."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local dev"
    }
  ],
...
  "paths": {
...
    "/api/brands": {
      "get": {
        "tags": ["Brands"],
        "operationId": "listBrand",
        "summary": "List brands",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "website": {
                            "type": "string"
                          },
                          "initials": {
                            "type": "string"
                          },
                          "tileClass": {
                            "type": "string"
                          },
                          "visibility": {
                            "type": "string",
                            "enum": ["Featured", "Standard", "Hidden"]
                          },
                          "status": {
                            "type": "string",
                            "enum": ["draft", "published", "archived"]
                          },
                          "productCount": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug",
                          "website",
                          "initials",
                          "tileClass",
                          "visibility",
                          "status",
                          "productCount",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  },
                  "required": ["data", "total", "page", "limit"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Brands"],
        "operationId": "createBrand",
        "summary": "Create brand",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "website": {
                    "type": "string",
                    "maxLength": 500,
                    "default": ""
                  },
                  "initials": {
                    "type": "string",
                    "maxLength": 4
                  },
                  "tileClass": {
                    "type": "string",
                    "default": "bg-brand-50 text-brand-600"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": ["Featured", "Standard", "Hidden"],
                    "default": "Standard"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published", "archived"],
                    "default": "draft"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "website": {
                      "type": "string"
                    },
                    "initials": {
                      "type": "string"
                    },
                    "tileClass": {
                      "type": "string"
                    },
                    "visibility": {
                      "type": "string",
                      "enum": ["Featured", "Standard", "Hidden"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["draft", "published", "archived"]
                    },
                    "productCount": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "website",
                    "initials",
                    "tileClass",
                    "visibility",
                    "status",
                    "productCount",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": ["error", "details"]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/brands/{id}": {
      "get": {
        "tags": ["Brands"],
        "operationId": "getBrand",
        "summary": "Get brand by ID",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Resource found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "website": {
                      "type": "string"
                    },
                    "initials": {
                      "type": "string"
                    },
                    "tileClass": {
                      "type": "string"
                    },
                    "visibility": {
                      "type": "string",
                      "enum": ["Featured", "Standard", "Hidden"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["draft", "published", "archived"]
                    },
                    "productCount": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "website",
                    "initials",
                    "tileClass",
                    "visibility",
                    "status",
                    "productCount",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Brands"],
        "operationId": "updateBrand",
        "summary": "Update brand",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "website": {
                    "type": "string",
                    "maxLength": 500,
                    "default": ""
                  },
                  "initials": {
                    "type": "string",
                    "maxLength": 4
                  },
                  "tileClass": {
                    "type": "string",
                    "default": "bg-brand-50 text-brand-600"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": ["Featured", "Standard", "Hidden"],
                    "default": "Standard"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published", "archived"],
                    "default": "draft"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "website": {
                      "type": "string"
                    },
                    "initials": {
                      "type": "string"
                    },
                    "tileClass": {
                      "type": "string"
                    },
                    "visibility": {
                      "type": "string",
                      "enum": ["Featured", "Standard", "Hidden"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["draft", "published", "archived"]
                    },
                    "productCount": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "website",
                    "initials",
                    "tileClass",
                    "visibility",
                    "status",
                    "productCount",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": ["error", "details"]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Brands"],
        "operationId": "deleteBrand",
        "summary": "Delete brand",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
  }
}